Lift up bump_fee_cpfp
